# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_ingest_upload_post": {
        "properties": {
          "declared_kpis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional JSON object of the entity's own declared KPIs, e.g. {\"mttr_minutes\": 45}. Taken as the submission states them, not as Orion's measurement \u2014 the distinction is what stops a declared figure from being read as a computed one.",
            "title": "Declared Kpis"
          },
          "entity_id": {
            "description": "Entity the submission is for",
            "title": "Entity Id",
            "type": "string"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "format": "binary",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "One or more source files. At least one is required.",
            "title": "Files"
          },
          "period_end": {
            "description": "Inclusive, YYYY-MM-DD",
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "description": "Inclusive, YYYY-MM-DD",
            "format": "date",
            "title": "Period Start",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "period_start",
          "period_end"
        ],
        "title": "Body_upload_api_v1_ingest_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionJobOut": {
        "additionalProperties": false,
        "properties": {
          "created_ts": {
            "title": "Created Ts",
            "type": "string"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished Ts"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "n_loaded": {
            "default": 0,
            "minimum": 0.0,
            "title": "N Loaded",
            "type": "integer"
          },
          "n_quarantined": {
            "default": 0,
            "minimum": 0.0,
            "title": "N Quarantined",
            "type": "integer"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started Ts"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          }
        },
        "required": [
          "job_id",
          "status",
          "created_ts"
        ],
        "title": "IngestionJobOut",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "complete",
          "failed",
          "quarantined"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "NotImplementedResponse": {
        "additionalProperties": false,
        "description": "Body returned by a route whose backing service is not built yet.\n\nThese routes are registered rather than omitted so that the OpenAPI document\nis the complete v1 surface: the frontend can be written against the whole API\nfrom day one and build only the parts that answer. The alternative \u2014\npublishing only the built half \u2014 means the contract changes under the\nfrontend halfway through the project, which is the failure mode a contract\nfreeze exists to prevent.\n\n`service` names what is missing and `phase` names when it lands, so nobody has\nto guess whether a 503 is a bug or a scheduled gap.",
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "phase": {
            "description": "Plan phase that delivers it",
            "title": "Phase",
            "type": "string"
          },
          "reason": {
            "default": "This endpoint is part of the frozen v1 contract but its backing ",
            "title": "Reason",
            "type": "string"
          },
          "service": {
            "description": "Service module not yet written",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "default": 503,
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "endpoint",
          "service",
          "phase"
        ],
        "title": "NotImplementedResponse",
        "type": "object"
      },
...
      "Page_IngestionJobOut_": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestionJobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "Page[IngestionJobOut]",
        "type": "object"
      },
...
      "PipelineStage": {
        "description": "The ten ingestion stages (plan \u00a74). Rejecting rows are never dropped.",
        "enum": [
          "receive",
          "verify",
          "detect",
          "map",
          "normalise",
          "validate",
          "quarantine",
          "pseudonymise",
          "load",
          "ledger"
        ],
        "title": "PipelineStage",
        "type": "string"
      },
...
      "UploadAccepted": {
        "additionalProperties": false,
        "description": "`POST /ingest/upload` \u2014 the job is queued, not done.\n\nIngestion runs in the background under the DuckDB single-writer lock, so\nthis is an acknowledgement, not a result.",
        "properties": {
          "accepted_files": {
            "items": {
              "type": "string"
            },
            "title": "Accepted Files",
            "type": "array"
          },
          "detected_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "detected_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Profile"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "total_bytes": {
            "minimum": 0.0,
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "entity_id",
          "status",
          "total_bytes",
          "message"
        ],
        "title": "UploadAccepted",
        "type": "object"
      },
      "UploadFormat": {
        "enum": [
          "csv",
          "tsv",
          "json",
          "ndjson",
          "parquet",
          "sqlite",
          "sql_dump",
          "xlsx"
        ],
        "title": "UploadFormat",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Supervisory Analytics Tool for SOC Assessment (SAT-SA) for NCIIPC.\n\nOutputs are **prioritised hypotheses with evidence**, never verdicts or\ncompliance grades. `Not assessable` is a first-class outcome, distinct from\n\"low risk\".\n\nFully offline: no cloud APIs, no CDN, no generative AI in the decision path.\n\n## Reading this contract\n\n`Not assessable` is a **value**, not an absence. Every score field is nullable\nand `null` means \"could not be assessed\" \u2014 never \"scored zero\". The frontend must\nnot coalesce `null` to `0`; a dimension with no data and a dimension with a\nmeasured score of zero are different claims.\n\n`POST /verdicts` records **a supervisor's own conclusion**, written by a human.\nNo Orion-derived payload in this API contains a verdict, a compliance grade, or a\npass/fail.\n\nRoutes marked `503` are part of this frozen contract but their backing service is\nnot written yet. The response body names the service and the phase that delivers\nit. They are listed rather than omitted so the surface does not change shape\nmid-project; `GET /api/v1` splits them into `live` and `pending`.",
    "title": "Orion (SAT-SA) Core Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingest/jobs": {
      "get": {
        "description": "One row per accepted upload. `quarantined` is a real outcome and not an error to hide: a submission that fails a security rule is parked with the reason recorded rather than dropped, so a missing file is always answerable.",
        "operationId": "list_jobs_api_v1_ingest_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "description": "queued, running, complete, failed, quarantined",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "queued, running, complete, failed, quarantined",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IngestionJobOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The ingestion queue",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingest/jobs/{job_id}": {
      "get": {
        "description": "The 202 from `POST /ingest/upload` hands back a job id and the load is asynchronous, because staging a multi-gigabyte file must not hold the request open. Poll this until `job_status` is terminal.",
        "operationId": "get_job_api_v1_ingest_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll one ingestion job",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingest/seed/{seed_id}": {
      "post": {
        "description": "In the frozen v1 contract. Returns 503 until the seed loader lands in Phase 19 \u2014 the SOCSim presets (OQ-8) are Dev 3's to decide, so the fixture set this route would load is not yet fixed.",
        "operationId": "load_seed_api_v1_ingest_seed__seed_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "seed_id",
            "required": true,
            "schema": {
              "title": "Seed Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Load a synthetic demonstration cohort (Phase 19)",
        "tags": [
          "ingestion"
        ],
        "x-orion": {
          "phase": "Phase 19 (2.19)",
          "service": "seed_loader",
          "status": "pending"
        }
      }
    },
    "/api/v1/ingest/upload": {
      "post": {
        "description": "Returns **202 with a job id**, not a result. The load runs under the DuckDB single-writer lock; poll `GET /ingest/jobs/{job_id}` for the outcome. A `202` here means the files were staged and a submission was registered \u2014 nothing has been parsed, validated or loaded yet.",
        "operationId": "upload_api_v1_ingest_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_ingest_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Upload rejected by the security checks"
          },
          "404": {
            "description": "No such entity"
          },
          "413": {
            "description": "Upload exceeds the configured size limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept a submission for an entity and queue it for loading",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  },
...
}